fix(dateRange): Use $scope.fromDate and $scope.toDate instead of dateRange
This removes dateRange object from gantt factory, in favor of more direct use of
$scope.fromDate and $scope.toDate. task-out-of-range default value has been changed to truncate.

Setting only from-date or to-date will works as expected. When no date is defined, gantt internal range is set for that all tasks to be visible. This range is available through $scope.gantt.from and $scope.gantt.to properties.

This also adds smalls optimizations and fix an issue related to auto-expand option.

Closes #169