fix(header): Show day header on hour view and week & month headers on day view
Depending on the current view scale the day or week & month is additionally shown in the header. Otherwise it is impossible to see from day numbers only what the current month or year is.

This commit fixes a regression introduced with d83f0e3dafd92b33dff5df9915d1967cfd5b7536