feat(layout): Use container available width when column-width is undefined
This is implemented by listening main div element width changes and updating column left/width values.

If column-width is undefined (default), width of each column will be updated for gantt to fill available width in container.

width property was dropped, and user should define width with css on the container.

Closes #120