refactor(movable): Extract moving and resizing of tasks to movable plugin
This extracts all tasks moving and resizing features to a new plugin.

It also make small changes to events ordering for plugins events and methods API to get registered before core.on.ready event is fired.

Attribute plugins are no longer supported, as the plugin scope may conflict with main gantt directive scope.

Plugins are children elements of gantt directive which is now a transcluding directive.

A ng-transclude node was added in default template to support plugins transclusion, and ng-cloack on main div to avoid blinking while initializing gantt.