fix(moveable): Task continues to scroll when moved to another row during scrolling
When moving a task to a new row while scrolling - the scrolling doesn't stop. This is because a new scope is generated when the task is moved to a new row but the old scope is still in the move-scroll-... cycle. The fix will set a flag when the scope is destroyed and the scroll cycle will stop if this flag is set.