fix(options): Boolean default options are not handled correctly
Boolean default options which are true are returned as undefined. Because if this an issue occurs with the side section. The side section has now a fixed defined height to simplify gantt resize calculations. Thus if no side plugin is loaded the side section should be hidden (default) otherwise its shown as empty.